Moved some functions in utils.py
